make test netcore builds only and coalesce into a single file
